% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNPN.R
\name{toNPN}
\alias{toNPN}
\title{Convert the GPS coordinate in WGS84 to NPN}
\usage{
toNPN(input.df, output = "han")
}
\arguments{
\item{input.df}{Input data frame containing longitude and latitude as columns}
}
\value{
NPN format (4 significant digits)
}
\description{
Convert the GPS coordinate in WGS84 to NPN
}
