% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toWGS.R
\name{toWGS}
\alias{toWGS}
\title{Convert NPN to the GPS coordinate in WGS84}
\usage{
toWGS(input.string)
}
\arguments{
\item{input.string}{Input character vector containing NPN properly spaced}
}
\value{
longitude and latitude
}
\description{
Convert NPN to the GPS coordinate in WGS84
}
